pyproject.toml
[project]
name = "cr01-timeseries-forecasting"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "einops",
    "transformers",
    "causal-conv1d>=1.4.0",
    "matplotlib>=3.10.3",
    "pandas>=2.3.1",
    "patool>=4.0.1",
    "pre-commit>=4.2.0",
    "reformer-pytorch>=1.4.4",
    "scikit-learn>=1.7.1",
    "sktime>=0.38.4",
    "torchinfo>=1.8.0",
    "tqdm>=4.67.1",
    "pywavelets>=1.9.0",
    "opt-einsum>=3.4.0",
    "mamba-ssm>=2.2.5",
    "gspread>=6.2.1",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "google-auth-httplib2>=0.2.0",
    "s5-pytorch>=0.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
main = "src.main:main"

# https://zenn.dev/turing_motors/articles/0f1a764d14f581
[tool.ruff]
exclude = [".git", ".mypy_cache", ".ruff_cache", ".venv", "third_party"]
line-length = 160
target-version = "py313"

[tool.ruff.lint]
fixable = ["ALL"]
unfixable = []
select = [
    "A",  # flake8-builtin
    "B",  # flake8-bugbear
    "E",  # pycodestyle error
    "F",  # Pyflakes
    "I",  # isort
    "N",  # pep8-naming
    "W",  # pycodestyle warning
    "PL",  # Pylint
    "UP",  # pyupgrade
]
ignore = [
    "B905",  # Zip-without-explicit-strict
    "E501",  # Line too long
    "F403",  # Unable to detect undefined names
    "N812",  # Lowercase imported as non-lowercase
    "N999",  # Invalid module name
    "PLR0912",  # Too many branches
    "PLR0913",  # Too many arguments in function definition
    "PLR2004",  # Magic value used in comparison
]

[tool.ruff.format]
quote-style = "double"
line-ending = "auto"

[tool.mypy]
allow_redefinition = true
allow_untyped_globals = false
check_untyped_defs = true
color_output = true
disallow_incomplete_defs = true
disallow_subclassing_any = false
disallow_untyped_calls = false
disallow_untyped_decorators = false
disallow_untyped_defs = true
error_summary = true
ignore_missing_imports = true
implicit_reexport = true
namespace_packages = true
no_implicit_optional = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
strict = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = false
exclude = [
  "mamba_ssm/",
  "mamba_ssm-*/",
]
